requirements.txt
# Scraping pipeline dependencies
openai>=1.40.0
supabase>=2.9.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
pyyaml>=6.0
